fixed wrong output sizes because pixel aspect ratio is ignored
Videos pixel aspect ratio is ignored when calculation output resolution.

Example:
Video with 720x576 [PAR 16:15] will be displayed with 768x576
720 * (16/15) = 768
Outputsize calculations should to be based on that resolution.

When resizing that video with "640x?" you expect a size of "640x480" as the source has an "4:3" aspect.
But what you get is a "640x512" sized videofile displayed at "682x512" because of the kept "16:15" pixel aspect ratio.

When resizing is based on the adjusted resolution, PAR will be "1:1" and the output will be the expected one.